Front-end and Back-end Files/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
numpy==1.26.4
gunicorn>=20.1.0
timm

